requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
numpy==2.0.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
scipy==1.13.1
simplejson==3.19.2
tb-mqtt-client==1.9.9
tb-paho-mqtt-client==1.6.3
urllib3==2.2.2
